refactor : 디렉토리명 reviews에서 review로 변경
