+++ src/main/java/sparta/ifour/movietalk/domain/review/dto/response/ReviewDetailResponseDto.java
@@ -1,36 +1,29 @@
-package sparta.ifour.movietalk.domain.reviews.dto.response;
+package sparta.ifour.movietalk.domain.review.dto.response;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import sparta.ifour.movietalk.domain.reviews.entity.Review;
-import sparta.ifour.movietalk.domain.reviews.entity.ReviewHashtag;
+import lombok.Getter;
+import sparta.ifour.movietalk.domain.comment.dto.response.CommentResponseDto;
+import sparta.ifour.movietalk.domain.review.entity.Review;
+import sparta.ifour.movietalk.domain.review.entity.ReviewHashtag;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class ReviewResponseDto {
+/**
+ * 상세 조회 관련 ResponseDto
+ */
+@Getter
+public class ReviewDetailResponseDto {
     private Long id;
-
     private String title;
-
     private String content;
-
     private Float ratingScore;
-
     private String movieName;
-
     private Integer countLikes;
-
     private List<ReviewHashtag> reviewHashtagList;
-
     private LocalDateTime createdAt;
+    private List<CommentResponseDto> commentList;
 
-
-    public ReviewResponseDto(Review review){
+    public ReviewDetailResponseDto(Review review) {
         this.id = review.getId();
         this.title = review.getTitle();
         this.content = review.getContent();
@@ -39,8 +32,9 @@ public class ReviewResponseDto {
         this.createdAt = review.getCreatedAt();
         this.countLikes = review.getLikeList().size();
         this.reviewHashtagList = review.getReviewHashtagList();
+        this.commentList = review.getCommentList().stream()
+                .map(CommentResponseDto::new)
+                .toList();
+
     }
-
-
-
 }
