+++ src/main/java/sparta/ifour/movietalk/domain/review/dto/response/ReviewResponseDto.java
@@ -1,30 +1,36 @@
-package sparta.ifour.movietalk.domain.reviews.dto.response;
+package sparta.ifour.movietalk.domain.review.dto.response;
 
-import lombok.Getter;
-import sparta.ifour.movietalk.domain.comment.dto.response.CommentCreateResponseDto;
-import sparta.ifour.movietalk.domain.comment.dto.response.CommentResponseDto;
-import sparta.ifour.movietalk.domain.reviews.entity.Review;
-import sparta.ifour.movietalk.domain.reviews.entity.ReviewHashtag;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import sparta.ifour.movietalk.domain.review.entity.Review;
+import sparta.ifour.movietalk.domain.review.entity.ReviewHashtag;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-/**
- * 상세 조회 관련 ResponseDto
- */
-@Getter
-public class ReviewDetailResponseDto {
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ReviewResponseDto {
     private Long id;
-    private String title;
-    private String content;
-    private Float ratingScore;
-    private String movieName;
-    private Integer countLikes;
-    private List<ReviewHashtag> reviewHashtagList;
-    private LocalDateTime createdAt;
-    private List<CommentResponseDto> commentList;
 
-    public ReviewDetailResponseDto(Review review) {
+    private String title;
+
+    private String content;
+
+    private Float ratingScore;
+
+    private String movieName;
+
+    private Integer countLikes;
+
+    private List<ReviewHashtag> reviewHashtagList;
+
+    private LocalDateTime createdAt;
+
+
+    public ReviewResponseDto(Review review){
         this.id = review.getId();
         this.title = review.getTitle();
         this.content = review.getContent();
@@ -33,9 +39,8 @@ public class ReviewDetailResponseDto {
         this.createdAt = review.getCreatedAt();
         this.countLikes = review.getLikeList().size();
         this.reviewHashtagList = review.getReviewHashtagList();
-        this.commentList = review.getCommentList().stream()
-                .map(CommentResponseDto::new)
-                .toList();
-
     }
+
+
+
 }
