+++ src/main/java/sparta/ifour/movietalk/domain/review/repository/LikeRepository.java
@@ -1,7 +1,7 @@
-package sparta.ifour.movietalk.domain.reviews.repository;
+package sparta.ifour.movietalk.domain.review.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sparta.ifour.movietalk.domain.reviews.entity.Like;
+import sparta.ifour.movietalk.domain.review.entity.Like;
 
 import java.util.Optional;
