+++ src/main/java/sparta/ifour/movietalk/domain/review/repository/ReviewHashTagRepository.java
@@ -1,7 +1,7 @@
-package sparta.ifour.movietalk.domain.reviews.repository;
+package sparta.ifour.movietalk.domain.review.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import sparta.ifour.movietalk.domain.reviews.entity.ReviewHashtag;
+import sparta.ifour.movietalk.domain.review.entity.ReviewHashtag;
 
 public interface ReviewHashTagRepository extends JpaRepository<ReviewHashtag,Long> {
 }
